fix(ci): python-full-coverage — align stale validation reason and split oversized tests
