+++ tests/unit/runtime/test_pr_monitor_runner_parts/test_pr_monitor_runner_part_002.py
@@ -1462,55 +1462,3 @@ async def test_pr_feedback_resolution_upsert_updates_same_comment_across_head_ch
     assert rows[0].source_operation_id == "op-new"
     assert rows[0].reason == "second monitor saw the inherited no-op verdict"
     assert fetched.head_sha == "new-head-after-repair-push"
-
-
-@pytest.mark.unit
-async def test_pr_feedback_resolution_body_change_creates_new_comment_identity(
-    factory: async_sessionmaker[AsyncSession],
-) -> None:
-    workspace_id = await seed_monitoring_workspace(factory)
-    async with factory() as session:
-        repo = PRFeedbackResolutionRepository(session)
-        await repo.record_resolution(
-            scm_provider="github",
-            repository_key="dimileeh/aira-web",
-            pull_request_key="42",
-            pull_request_url="https://github.com/dimileeh/aira-web/pull/42",
-            head_sha="old-head",
-            feedback_kind="review_comment",
-            feedback_id="issue:4391271818",
-            feedback_body="old body",
-            feedback_author="chatgpt-codex-connector[bot]",
-            feedback_url="https://github.example/comment/4391271818",
-            verdict="false_positive",
-            reason="old comment body",
-            source_workspace_id=workspace_id,
-        )
-        await repo.record_resolution(
-            scm_provider="github",
-            repository_key="dimileeh/aira-web",
-            pull_request_key="42",
-            pull_request_url="https://github.com/dimileeh/aira-web/pull/42",
-            head_sha="new-head",
-            feedback_kind="review_comment",
-            feedback_id="issue:4391271818",
-            feedback_body="new body with new actionable content",
-            feedback_author="chatgpt-codex-connector[bot]",
-            feedback_url="https://github.example/comment/4391271818",
-            verdict="defer",
-            reason="body changed, so the monitor must re-evaluate it",
-            source_workspace_id=workspace_id,
-        )
-        await session.commit()
-
-        rows = await repo.list_for_pr(
-            scm_provider="github",
-            repository_key="dimileeh/aira-web",
-            pull_request_key="42",
-        )
-
-    assert len(rows) == 2
-    assert {row.reason for row in rows} == {
-        "old comment body",
-        "body changed, so the monitor must re-evaluate it",
-    }
