+++ tests/unit/service/test_workspace_response_parts/test_workspace_response_part_001.py
@@ -21,9 +21,6 @@ from awf.service import workspaces_create as workspaces_create_service
 from awf.service import workspaces_response as workspaces_response_service
 from awf.service import workspaces_retry as workspaces_retry_service
 from awf.service.validation_observability import (
-    _profile_requested_validation_tier,
-    _summary_freshness_and_reason,
-    latest_merge_candidate,
     validation_freshness_summary,
 )
 from awf.service.workspaces import workspace_failure_details_payload, workspace_response
@@ -1475,61 +1472,3 @@ def test_workspace_validation_summary_uses_latest_successful_rebase() -> None:
     assert summary.reason_code == "validation_insufficient_tier"
     assert summary.latest_validation is not None
     assert summary.latest_validation.freshness_status == "fresh"
-
-
-@pytest.mark.unit
-def test_latest_merge_candidate_ignores_candidates_with_missing_status() -> None:
-    newer_missing_status = SimpleNamespace(
-        id="mc_missing_status",
-        updated_at=datetime(2026, 4, 27, 16, 0, tzinfo=UTC),
-    )
-    older_open = SimpleNamespace(
-        id="mc_open",
-        status="open",
-        updated_at=datetime(2026, 4, 27, 15, 0, tzinfo=UTC),
-    )
-    workspace = SimpleNamespace(merge_candidates=[newer_missing_status, older_open])
-
-    assert latest_merge_candidate(workspace) is older_open  # type: ignore[arg-type]
-
-
-@pytest.mark.unit
-def test_validation_summary_propagates_collection_access_errors() -> None:
-    class WorkspaceWithBrokenOperations:
-        task_class = None
-        resolved_profile = None
-        monitor_last_commit_sha = None
-
-        @property
-        def operations(self) -> list[object]:
-            raise RuntimeError("relationship failed")
-
-    with pytest.raises(RuntimeError, match="relationship failed"):
-        validation_freshness_summary(
-            WorkspaceWithBrokenOperations(),  # type: ignore[arg-type]
-            [],
-        )
-
-
-@pytest.mark.unit
-def test_validation_freshness_reason_defaults_when_latest_run_has_no_reason() -> None:
-    freshness, reason = _summary_freshness_and_reason(
-        required_tier=1,
-        latest_satisfied_tier=1,
-        latest_validation=SimpleNamespace(
-            freshness_status="fresh",
-            freshness_reason_code=None,
-        ),  # type: ignore[arg-type]
-    )
-
-    assert freshness == "fresh"
-    assert reason == "validation_target_unknown"
-
-
-@pytest.mark.unit
-def test_profile_requested_validation_tier_ignores_non_integer_profile_value() -> None:
-    workspace = SimpleNamespace(
-        resolved_profile={"validation": {"requested_tier": "3"}},
-    )
-
-    assert _profile_requested_validation_tier(workspace) == 1  # type: ignore[arg-type]
